beachvolleyball/requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.4.0
Django==2.1.7
django-extensions==2.1.6
djangorestframework==3.9.2
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
psycopg2-binary
django-environ=0.4.5
djangorestframework-xml=1.4.0
defusedxml=0.5.0
